Move the chowning / chgrping to CreateNixTree, and prevent C.N.T. from erroring on UID mismatch.
